# go_web/desafio-go-web/docs/swagger.yaml
basePath: /
definitions:
  domain.AverageTicketsResponse:
    properties:
      average:
        type: number
    type: object
  domain.TotalTicketsResponse:
    properties:
      totalTickets:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Marina Rocha Lhamby
  title: Desafio Go Web Docs
  version: 1.0.0
paths:
  /tickets:
    get:
      consumes:
      - application/json
      description: Fetch total number of tickets by period
      parameters:
      - description: início da manhã, manhã, tarde ou noite
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/domain.TotalTicketsResponse'
            type: object
        "400":
          description: error
          schema:
            type: string
      summary: Fetch total number of tickets by period
      tags:
      - tickets
  /tickets/destination/{destination}:
    get:
      consumes:
      - application/json
      description: Fetch number of tickets by destination
      parameters:
      - description: Destination
        in: path
        name: destination
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/domain.TotalTicketsResponse'
            type: object
      summary: Fetch number of tickets by destination
      tags:
      - tickets
  /tickets/destination/average/{destination}:
    get:
      consumes:
      - application/json
      description: Fetch average number of tickets by destination
      parameters:
      - description: Destination
        in: path
        name: destination
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/domain.AverageTicketsResponse'
            type: object
      summary: Fetch average number of tickets by destination
      tags:
      - tickets
swagger: "2.0"